refactor(jobdb_orm): adjust jobdb_orm directory
