+++ store/jobdb/db_schema.go
@@ -1,4 +1,4 @@
-package database
+package jobdb
 
 import (
 	"fmt"
