+++ store/jobdb/db_schema_test.go
@@ -1,4 +1,4 @@
-package database
+package jobdb
 
 import (
 	"testing"
